+++ xcloud-devops-ci/xcloud-devops-ci-core/src/main/java/com/wl4g/devops/ci/core/PipelineManager.java
@@ -59,7 +59,7 @@ public abstract interface PipelineManager {
 	 * @param size
 	 * @return
 	 */
-	ReadResult logfile(Integer taskHisId, Long startPos, Integer size);
+	ReadResult logfile(Long taskHisId, Long startPos, Integer size);
 
 	/**
 	 * Reader pipeline task detail deploying logs.
@@ -70,6 +70,6 @@ public abstract interface PipelineManager {
 	 * @param size
 	 * @return
 	 */
-	ReadResult logDetailFile(Integer taskHisId, Integer instanceId, Long startPos, Integer size);
+	ReadResult logDetailFile(Long taskHisId, Long instanceId, Long startPos, Integer size);
 
 }
