+++ xcloud-devops-ci/xcloud-devops-ci-core/src/main/java/com/wl4g/devops/ci/pipeline/pcm/PcmPipelineProvider.java
@@ -23,6 +23,6 @@ import com.wl4g.components.core.bean.ci.PipeHistoryPcm;
  */
 public interface PcmPipelineProvider {
 
-	void createIssues(Integer pcmId, PipeHistoryPcm pipeHistoryPcm);
+	void createIssues(Long pcmId, PipeHistoryPcm pipeHistoryPcm);
 
 }
