+++ xcloud-devops-ci/xcloud-devops-ci-core/src/main/java/com/wl4g/devops/ci/service/impl/PcmServcieImpl.java
@@ -79,7 +79,7 @@ public class PcmServcieImpl implements PcmService {
 	}
 
 	@Override
-	public void del(Integer id) {
+	public void del(Long id) {
 		Pcm pcm = new Pcm();
 		pcm.setId(id);
 		pcm.setDelFlag(BaseBean.DEL_FLAG_DELETE);
@@ -87,7 +87,7 @@ public class PcmServcieImpl implements PcmService {
 	}
 
 	@Override
-	public Pcm detail(Integer id) {
+	public Pcm detail(Long id) {
 		return pcmDao.selectByPrimaryKey(id);
 	}
 
@@ -97,7 +97,7 @@ public class PcmServcieImpl implements PcmService {
 	}
 
 	@Override
-	public List<SelectionModel> getUsers(Integer pcmId) {
+	public List<SelectionModel> getUsers(Long pcmId) {
 		Pcm pcm = getPcm(pcmId);
 		if (Objects.isNull(pcm)) {
 			return null;
@@ -106,7 +106,7 @@ public class PcmServcieImpl implements PcmService {
 	}
 
 	@Override
-	public List<SelectionModel> getProjects(Integer pcmId) {
+	public List<SelectionModel> getProjects(Long pcmId) {
 		Pcm pcm = getPcm(pcmId);
 		if (Objects.isNull(pcm)) {
 			return null;
@@ -115,7 +115,7 @@ public class PcmServcieImpl implements PcmService {
 	}
 
 	@Override
-	public List<SelectionModel> getIssues(Integer pcmId, String userId, String projectId, String search) {
+	public List<SelectionModel> getIssues(Long pcmId, String userId, String projectId, String search) {
 		Pcm pcm = getPcm(pcmId);
 		if (Objects.isNull(pcm)) {
 			return null;
@@ -124,24 +124,24 @@ public class PcmServcieImpl implements PcmService {
 	}
 
 	@Override
-	public List<SelectionModel> getProjectsByPcmId(Integer pcmId) {
+	public List<SelectionModel> getProjectsByPcmId(Long pcmId) {
 		Pcm pcm = getPcm(pcmId);
 		return pcmOperator.forOperator(pcm.getProviderKind()).getProjects(pcm);
 	}
 
 	@Override
-	public List<SelectionModel> getTrackers(Integer pcmId) {
+	public List<SelectionModel> getTrackers(Long pcmId) {
 		Pcm pcm = getPcm(pcmId);
 		return pcmOperator.forOperator(pcm.getProviderKind()).getTracker(pcm);
 	}
 
 	@Override
-	public List<SelectionModel> getStatuses(Integer pcmId) {
+	public List<SelectionModel> getStatuses(Long pcmId) {
 		Pcm pcm = getPcm(pcmId);
 		return pcmOperator.forOperator(pcm.getProviderKind()).getStatuses(pcm);
 	}
 
-	private Pcm getPcm(Integer pcmId) {
+	private Pcm getPcm(Long pcmId) {
 		Pcm pcm = pcmDao.selectByPrimaryKey(pcmId);
 		if (Objects.isNull(pcm)) {
 			Page<Pcm> list = pcmDao.list(getRequestOrganizationCodes(), null, null, null);
@@ -154,13 +154,13 @@ public class PcmServcieImpl implements PcmService {
 	}
 
 	@Override
-	public List<SelectionModel> getPriorities(Integer pcmId) {
+	public List<SelectionModel> getPriorities(Long pcmId) {
 		Pcm pcm = pcmDao.selectByPrimaryKey(pcmId);
 		return pcmOperator.forOperator(pcm.getProviderKind()).getPriorities(pcm);
 	}
 
 	@Override
-	public void createIssues(Integer pcmId, PipeHistoryPcm pipeHistoryPcm) {
+	public void createIssues(Long pcmId, PipeHistoryPcm pipeHistoryPcm) {
 		Pcm pcm = pcmDao.selectByPrimaryKey(pcmId);
 		pcmOperator.forOperator(pcm.getProviderKind()).createIssues(pcm, pipeHistoryPcm);
